requirements.txt
PyYAML
Pillow
python-dotenv
